Make Concourse resilient to worker pod restarts.
Add time between worker stop and start. A shell is used to start the worker and sleep after the process terminates. SIGTERM is not propagated to the concourse process as a result, but this is actually desired for the shutdown process, below.

Gracefully de-register a worker using `retire-worker` instead of `land-worker` in the `preStop` hook. This tells concourse that the worker and its container volumes will not be coming back, which is the case when worker pods restart. Because SIGTERM doesn't reach the concourse process, `retire-worker` is the only mechanism used for graceful shutdown. If all containers are cleaned up within `worker.terminationGracePeriodSeconds`, the concourse process terminates on its own and the pod finishes terminating, otherwise the pod's container is killed. In either case, the worker disappears on ATC's end. See https://concourse.ci/worker-internals.html for more info.

Fixes https://github.com/kubernetes/charts/issues/2103