[stable/datadog] bugfix cluster-agent service name
The current default is not to not configure the cluster-agent deployment
with the service name, this causes the cluster check to fail.

The bug will be triggered when cluster-agent is enabled and fullname is not
datadog. The cluster agent pods will be attempting to connect to the
wrong service name.

Currently the cluster-agent service name is only passed to the
cluster-agent deployment when collectEvents is true. The change is to
always configure the service name in the cluster-agent deployment.

Signed-off-by: Josh Switnicki <switj@autonomic.ai>